backend/requirements.txt
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
requests==2.31.0
beautifulsoup4==4.12.2
aiosqlite==0.19.0
httpx==0.26.0
